+++ acceptance/brutil/br-core-conf/topology.json
@@ -57,6 +57,15 @@
           "ISD_AS": "1-ff00:0:4",
           "RemoteOverlay": { "OverlayPort": 40000, "Addr": "192.168.14.5" },
           "Bandwidth": 1000
+        },
+        "1502": {
+          "LinkTo": "CHILD",
+          "MTU": 1472,
+          "Overlay": "UDP/IPv4",
+          "PublicOverlay": { "OverlayPort": 50000, "Addr": "192.168.15.4" },
+          "ISD_AS": "1-ff00:0:5",
+          "RemoteOverlay": { "OverlayPort": 40000, "Addr": "192.168.15.5" },
+          "Bandwidth": 1000
         }
       }
     },
