+++ acceptance/brutil/util.sh
@@ -1,4 +1,6 @@
-#!/bin/bash
+# This is a base file included/sourced by each border router acceptance test
+
+# Following are the specific setup functions for border acceptance tests
 
 # syntax: create_veth <netns> <host_if_name> <container_if_name> <container_ip_addr> <neigh_ips>
 create_veth() {
@@ -18,7 +20,7 @@ create_veth() {
         sudo ip netns exec $NS ip neigh add $ip lladdr f0:0d:ca:fe:be:ef nud permanent dev $VETH_CONTAINER
     done
     sudo ip netns exec $NS ip link set $VETH_CONTAINER up
-	MAC=$(sudo ip netns exec $NS cat /sys/class/net/${VETH_CONTAINER}/address)
+    MAC=$(sudo ip netns exec $NS cat /sys/class/net/${VETH_CONTAINER}/address)
     echo "$VETH_HOST $VETH_CONTAINER $MAC"
 }
 
@@ -27,17 +29,17 @@ get_docker_ns_path() {
 }
 
 get_docker_ns() {
-	NS_path=$(get_docker_ns_path)
-	echo $(basename $NS_path)
+    NS_path=$(get_docker_ns_path)
+    echo $(basename $NS_path)
 }
 
 set_docker_ns_link() {
-	NS_path=$(get_docker_ns_path)
+    NS_path=$(get_docker_ns_path)
     sudo mkdir -p /var/run/netns && sudo ln -fs $NS_path -t /var/run/netns
 }
 
 rm_docker_ns_link() {
-	NS=$(get_docker_ns)
+    NS=$(get_docker_ns)
     sudo rm -f /var/run/netns/$NS
 }
 
@@ -46,3 +48,4 @@ delete_veth() {
         sudo ip link set $dev down && sudo ip link del $dev || true
     done
 }
+
