+++ acceptance/core_brB_acceptance/test
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-BRID=brA
+BRID=core-brB
 TEST_NAME="${BRID}_acceptance"
 PROGRAM=$(basename "$0")
 COMMAND="$1"
@@ -12,15 +12,15 @@ BRUTIL=acceptance/brutil
 
 # This function is called from test_setup
 set_veths() {
-    create_veth $NS veth0_root ifid_local 192.168.0.11/24 \
-        192.168.0.12 192.168.0.13 192.168.0.14 192.168.0.51 192.168.0.61 192.168.0.71
-    create_veth $NS veth1_root ifid_1201 192.168.12.2/31 192.168.12.3
+    create_veth $NS veth0_root ifid_local 192.168.0.12/24 \
+        192.168.0.11 192.168.0.13 192.168.0.51 192.168.0.61 192.168.0.71
+    create_veth $NS veth1_root ifid_1401 192.168.14.2/31 192.168.14.3
 }
 
 test_run() {
     set -e
     bin/braccept -borderID "$BRID" -devInfoFilePath "$DEVINFO_FN" \
-        -keysDirPath "$BRUTIL/br-conf/keys" $@
+        -keysDirPath "$BRUTIL/br-core-conf/keys" $@
 }
 
 # This function is called from test_teardown
