+++ go/border/braccept/pkti/overlay.go
@@ -1,0 +1,102 @@
+package pkti
+
+import (
+	"fmt"
+	"net"
+
+	"github.com/google/gopacket"
+	"github.com/google/gopacket/layers"
+)
+
+// OverlayLayers is the interface used to generate overlays.
+type OverlayLayers interface {
+	ToLayers() []gopacket.SerializableLayer
+	Check(l []gopacket.Layer) ([]gopacket.Layer, error)
+}
+
+// OverlayIP4UDP implementes the IPv4/UDP overlay
+type OverlayIP4UDP struct {
+	SrcAddr string
+	SrcPort uint16
+	DstAddr string
+	DstPort uint16
+}
+
+func (o *OverlayIP4UDP) ToLayers() (l []gopacket.SerializableLayer) {
+	srcIP := net.ParseIP(o.SrcAddr)
+	dstIP := net.ParseIP(o.DstAddr)
+	var nl gopacket.NetworkLayer
+	ip := &layers.IPv4{
+		Version:  4,
+		IHL:      5,
+		TTL:      64,
+		Protocol: layers.IPProtocolUDP,
+		SrcIP:    srcIP,
+		DstIP:    dstIP,
+	}
+	nl = ip
+	l = append(l, ip)
+	udp := &layers.UDP{
+		SrcPort: layers.UDPPort(o.SrcPort),
+		DstPort: layers.UDPPort(o.DstPort),
+	}
+	l = append(l, udp)
+	udp.SetNetworkLayerForChecksum(nl)
+	return l
+}
+
+func (o *OverlayIP4UDP) Check(l []gopacket.Layer) ([]gopacket.Layer, error) {
+	overlayLayers := o.ToLayers()
+	for i := range overlayLayers {
+		if err := compareLayer(l[i], overlayLayers[i]); err != nil {
+			return l[i:], err
+		}
+	}
+	return l[len(overlayLayers):], nil
+}
+
+func compareLayer(act gopacket.Layer, exp gopacket.SerializableLayer) error {
+	if exp.LayerType() != act.LayerType() {
+		return fmt.Errorf("Wrong Layer Type, expected %s, actuaact %s",
+			exp.LayerType(), act.LayerType())
+	}
+	switch exp.LayerType() {
+	case layers.LayerTypeIPv4:
+		return compareIP4Layer(exp.(*layers.IPv4), act.(*layers.IPv4))
+	case layers.LayerTypeIPv6:
+		return compareIP6Layer(exp.(*layers.IPv6), act.(*layers.IPv6))
+	case layers.LayerTypeUDP:
+		return compareUDPLayer(exp.(*layers.UDP), act.(*layers.UDP))
+	}
+	return fmt.Errorf("Unknown layer %s", exp.LayerType())
+}
+
+func compareIP4Layer(act, exp *layers.IPv4) error {
+	return compareIPLayer(act.SrcIP, act.DstIP, exp.SrcIP, exp.DstIP)
+}
+
+func compareIP6Layer(act, exp *layers.IPv6) error {
+	return compareIPLayer(act.SrcIP, act.DstIP, exp.SrcIP, exp.DstIP)
+}
+
+func compareIPLayer(actSrcIP, actDstIP, expSrcIP, expDstIP net.IP) error {
+	if !actSrcIP.Equal(expSrcIP) {
+		return fmt.Errorf("Wrong Source IP, expected %s, actual %s", expSrcIP, actSrcIP)
+	}
+	if !actDstIP.Equal(expDstIP) {
+		return fmt.Errorf("Wrong Destination IP, expected %s, actual %s", expDstIP, actDstIP)
+	}
+	return nil
+}
+
+func compareUDPLayer(act, exp *layers.UDP) error {
+	if act.SrcPort != exp.SrcPort {
+		return fmt.Errorf("Wrong UDP Source Port, expected %s, actual %s",
+			exp.SrcPort, act.SrcPort)
+	}
+	if act.DstPort != exp.DstPort {
+		return fmt.Errorf("Wrong UDP Destination Port, expected %s, actual %s",
+			exp.DstPort, act.DstPort)
+	}
+	return nil
+}
